update borad UI
