+++ src/edu/miu/cs/cs525/reversi/view/BoardUI.fxml
@@ -9,7 +9,8 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.shape.Box?>
-<BorderPane prefHeight="540"
+<?import edu.miu.cs.cs525.reversi.model.Tile?>
+<BorderPane prefHeight="600"
             style="-fx-background-color: darkslategray"
             prefWidth="680"
             xmlns="http://javafx.com/javafx/8.0.111"
@@ -83,9 +84,8 @@
                         </graphic>
                     </Button>
                 </HBox>
-                <BorderPane styleClass="board" prefHeight="540" prefWidth="540" layoutY="25">
-
-                </BorderPane>
+                <StackPane fx:id="stackPane" styleClass="board" layoutY="25" prefHeight="540" prefWidth="540">
+                </StackPane>
             </children>
         </Pane>
     </center>
